Add TLS support for Elasticsearch connections (#34)
Adds TLS support for purposes of connecting to a secured ElasticSearch
instance. Testing was done with ElasticSearch as deployed via the
ElasticSearch Operator available at
https://github.com/openshift/elasticsearch-operator.

If TlsServerName is configured for the Smart Gateway, then that is
used as part of the connection setup to validate the server name as
defined in the connection setup. If TlsServerName is not defined then
InsecureSkipVerify is set to a value of true which results in skipping
that connection setup validation.

Co-Authored-By: Martin Magr <mmagr@redhat.com>
Co-Authored-By: Leif Madsen <lmadsen@redhat.com>

Closes #32