Don't act like a bitcoin miner in Events mode
When running a Smart Gateway in Events mode, the CPU would run at a high
level when idling (not processing events). After some profiling with pprof
it was identified that the CPU was hanging in a case loop with a default that
wasn't doing any processing (thus not blocking).

Thanks to Chris Sibbitt and Aaron Smith for the help pointing out issue.

Closes #54