Define travis-ci as our goveralls service
Update the goveralls service type to travis-ci since that's the CI
platform we're running on top of.

Also had to utilize a new repository that holds git 2.x for CentOS 7
so that the goveralls command (which recently had a chance to the
gitinfo.go code causing a breaking change on git 1.8) could complete
successfully.

Disabled ipv6 by force an ipv4 connection in yum to work around
installation issues with the new repository.