Base logic for event handlers
This patch adds basic logic for event handlers. Currently it's not plugin based,
but the implementation is prepared for that.

This patch also adds first handler included in the SG codebase to not change
build process now. Thi handler splits the check-container-healthcheck result
event from collectd-sensubility and saves it to ES database as multiple events.