Base logic for event handlers (#95)
* Base logic for event handlers

This patch adds basic logic for event handlers. Currently it's not plugin based,
but the implementation is prepared for that.

This patch also adds first handler included in the SG codebase to not change
build process now. Thi handler splits the check-container-healthcheck result
event from collectd-sensubility and saves it to ES database as multiple events.

* Fix handler issues

    - leaving output in json for better querying
    - saving all output items, not just first one
    - fixed naming typo

* Fix staticcheck issue

* Fix lint issue

* refactor handler function (#96)

Co-authored-by: pleimer <pleimer@redhat.com>

Co-authored-by: pleimer <41337120+pleimer@users.noreply.github.com>
Co-authored-by: pleimer <pleimer@redhat.com>